// src/app/api/common/types.ts
export enum VERIFIED_CURRENCY {
  // CREDITS = "CREDITS",
  USDC = "USDC",
  SOL = "SOL",
  BONK = "BONK",
  SEND = "SEND",
}
// ...
export enum CLUSTER_TYPES {
  DEVNET = "devnet",
  MAINNET = "mainnet",
}

export interface IWeb3Participate {
  account: PublicKey;
  playerId: BN;
  challengeId: BN;
  amount: BN;
  currency?: VERIFIED_CURRENCY;
  onchainParticipateType: ONCHAIN_PARTICIPATE_TYPE;
}

export enum ONCHAIN_PARTICIPATE_TYPE {
  JOIN_CHALLENGE = "joinChallenge",
  SIDE_BET = "sideBet",
}

export interface IGetTxObject extends IWeb3Participate {
  userPublicKey: PublicKey;
  escrowTokenAccount: PublicKey;
  userTokenAccount: PublicKey;
  program: Program;
  cluster: keyof typeof ONCHAIN_CONFIG;
}
// ...
export interface IFindMyCatById {
  Name: number;
  Media: string;
  Currency: VERIFIED_CURRENCY;
  Wager: number;
}